requirements.txt
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
Flask==3.1.0
requests==2.32.3
schedule==1.2.0